requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.4.2
python-multipart>=0.0.9
gradio==4.19.2
httpx==0.26.0
